secim-ekonomisi-starter/secim-ekonomisi/requirements.txt
pandas
pyyaml
requests
matplotlib
pyarrow
